The user can add, edit and delete an album

+++ app/src/main/java/com/example/android/youtubemusicplayer/database/Playlist.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "playlists_table")
 data class Playlist (
     @PrimaryKey(autoGenerate = true)
-    val playlistId: Int = 0,
+    val playlistId: Long = 0,
 
     var name: String = ""
 )
